requirements.txt
requests
cloudflare>3
pyyaml
pythonping
loguru